Remove SDL fixed-width types from graphics/network/event code. (#305)
This gets rid of the vast majority of uses of SDL fixed-width integer
types from MZX, leaving only a few in SDL-specific places, in the audio
subsystem, and in platform.h. This should further reduce the number of
instances of includes of SDL headers via removing the platform.h
include from event.h and graphics.h. SDL headers pull in a bunch of
other headers and potentially slow down builds in VMs and older systems
significantly.

I also took this opportunity to clean up some of the graphics code,
particularly the drawing functions (all "linear" drawing functions have
been removed). Instances of fixed-width types that seemed important or
at least useful now use the stdint.h types instead of SDL types, but
most previous uses were just replaced with int, unsigned int, or size_t.

Other changes of note:

* Luma functions in graphics.c are now __editor_maybe_static.
* Uses of FILE/stdio in graphics.c and pal_ed.c have been replaced with
  vio functions.
* Mouse button constants are no longer dependent on the existence of
  SDL and the SDL version. event_sdl.c now translates SDL buttons to
  MZX buttons.
* Removed some unused graphics.h and platform.h includes.
* Removed dump_char, which has been replaced with different functions
  that do a better job at what they're intended for than dump_char did.
* The classic sprite bleedthrough effect is now handled by a single
  graphics.c function instead of two, simplifying the code.
* All graphics.c drawing functions now require X/Y instead of a linear
  position, meaning layer offset conversion can be done faster. Layer
  offset conversion is now skipped * for layers at 0,0 with the same
  width as text_video.
* Added set_colors_function typedef to render.h to clean up
  declarations/usage of the set_colors arrays, render_graph8, and
  render_graph16.
* SMZX indices should now work correctly in the GP2X renderer.

+++ src/network/sha256.c
@@ -35,7 +35,7 @@
 #define sig0(x)    (S(x, 7) ^ S(x,18) ^ R(x, 3))
 #define sig1(x)    (S(x,17) ^ S(x,19) ^ R(x,10))
 
-static const Uint32 K[] =
+static const uint32_t K[] =
 {
   0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5,
   0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
@@ -55,7 +55,7 @@ static const Uint32 K[] =
   0x90befffa, 0xa4506ceb, 0xbef9a3f7, 0xc67178f2
 };
 
-static const Uint32 H_initial[] = {
+static const uint32_t H_initial[] = {
   0x6a09e667, 0xbb67ae85, 0x3c6ef372, 0xa54ff53a,
   0x510e527f, 0x9b05688c, 0x1f83d9ab, 0x5be0cd19
 };
@@ -64,16 +64,16 @@ static const Uint32 H_initial[] = {
 
 static void convert_to_bigendian(void *data, int len)
 {
-  Uint32 *data_as_words = (Uint32 *)data;
-  Uint8 *data_as_bytes;
-  Uint32 temp;
-  Uint8 *temp_as_bytes = (Uint8 *)&temp;
+  uint32_t *data_as_words = (uint32_t *)data;
+  uint8_t *data_as_bytes;
+  uint32_t temp;
+  uint8_t *temp_as_bytes = (uint8_t *)&temp;
   int i;
 
   for(i = 0; i < len / 4; i++)
   {
     temp = data_as_words[i];
-    data_as_bytes = (Uint8 *)&data_as_words[i];
+    data_as_bytes = (uint8_t *)&data_as_words[i];
 
     data_as_bytes[0] = temp_as_bytes[3];
     data_as_bytes[1] = temp_as_bytes[2];
@@ -90,16 +90,16 @@ static inline void convert_to_bigendian(void *data, int len) { }
 
 static void SHA256_transform(struct SHA256_ctx *ctx)
  {
-  Uint32 A = ctx->H[0];
-  Uint32 B = ctx->H[1];
-  Uint32 C = ctx->H[2];
-  Uint32 D = ctx->H[3];
-  Uint32 E = ctx->H[4];
-  Uint32 F = ctx->H[5];
-  Uint32 G = ctx->H[6];
-  Uint32 H = ctx->H[7];
-  Uint32 T1, T2;
-  Uint32 W[64];
+  uint32_t A = ctx->H[0];
+  uint32_t B = ctx->H[1];
+  uint32_t C = ctx->H[2];
+  uint32_t D = ctx->H[3];
+  uint32_t E = ctx->H[4];
+  uint32_t F = ctx->H[5];
+  uint32_t G = ctx->H[6];
+  uint32_t H = ctx->H[7];
+  uint32_t T1, T2;
+  uint32_t W[64];
   int t;
 
   memcpy(W, ctx->M, 64);
@@ -133,17 +133,17 @@ static void SHA256_transform(struct SHA256_ctx *ctx)
 
 void SHA256_init(struct SHA256_ctx *ctx)
 {
-  memcpy(ctx->H, H_initial, 8 * sizeof(Uint32));
+  memcpy(ctx->H, H_initial, 8 * sizeof(uint32_t));
   ctx->lbits = 0;
   ctx->hbits = 0;
   ctx->mlen = 0;
 }
 
-void SHA256_update(struct SHA256_ctx *ctx, const void *vdata, Uint32 data_len)
+void SHA256_update(struct SHA256_ctx *ctx, const void *vdata, size_t data_len)
 {
-  const Uint8 *data = (const Uint8 *)vdata;
-  Uint32 low_bits;
-  Uint32 use;
+  const uint8_t *data = (const uint8_t *)vdata;
+  uint32_t low_bits;
+  size_t use;
 
   /* convert data_len to bits and add to the 64 bit word
    * formed by lbits and hbits
@@ -158,8 +158,8 @@ void SHA256_update(struct SHA256_ctx *ctx, const void *vdata, Uint32 data_len)
 
   /* deal with first block */
 
-  use = MIN((Uint32)(64 - ctx->mlen), data_len);
-  memcpy(ctx->M + ctx->mlen, data, use );
+  use = MIN((size_t)(64 - ctx->mlen), data_len);
+  memcpy(ctx->M + ctx->mlen, data, use);
   ctx->mlen += use;
   data_len -= use;
   data += use;
@@ -205,14 +205,14 @@ int main(void)
   int i;
 
   const char *test1 = "abc";
-  const Uint32 result1[8] = {
+  const uint32_t result1[8] = {
     0xba7816bf, 0x8f01cfea, 0x414140de, 0x5dae2223,
     0xb00361a3, 0x96177a9c, 0xb410ff61, 0xf20015ad
   };
 
   const char *test2 =
     "abcdbcdecdefdefgefghfghighijhijkijkljklmklmnlmnomnopnopq";
-  const Uint32 result2[8] = {
+  const uint32_t result2[8] = {
     0x248d6a61, 0xd20638b8, 0xe5c02693, 0x0c3e6039,
     0xa33ce459, 0x64ff2167, 0xf6ecedd4, 0x19db06c1
   };
